# .migration-backup/lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * Mr Farrukh Hair Saloon API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Appointment {
  id: number;
  name: string;
  phone: string;
  /** @nullable */
  email?: string | null;
  service: string;
  /** @nullable */
  stylist?: string | null;
  date: string;
  time: string;
  /** @nullable */
  notes?: string | null;
  status: string;
  createdAt: string;
}

export interface AppointmentInput {
  name: string;
  phone: string;
  email?: string;
  service: string;
  stylist?: string;
  date: string;
  time: string;
  notes?: string;
}

export interface AppointmentUpdate {
  status?: string;
  notes?: string;
}

export interface Service {
  id: number;
  name: string;
  category: string;
  description: string;
  /** @nullable */
  imageUrl?: string | null;
  /** @nullable */
  duration?: string | null;
  featured?: boolean;
}

export interface ServiceInput {
  name: string;
  category: string;
  description: string;
  imageUrl?: string;
  duration?: string;
  featured?: boolean;
}

export interface ServiceUpdate {
  name?: string;
  category?: string;
  description?: string;
  imageUrl?: string;
  duration?: string;
  featured?: boolean;
}

export interface PricingItem {
  id: number;
  name: string;
  category: string;
  price: number;
  /** @nullable */
  unit?: string | null;
  popular?: boolean;
}

export interface PricingItemInput {
  name: string;
  category: string;
  price: number;
  unit?: string;
  popular?: boolean;
}

export interface PricingItemUpdate {
  name?: string;
  category?: string;
  price?: number;
  unit?: string;
  popular?: boolean;
}

export interface GalleryItem {
  id: number;
  imageUrl: string;
  category: string;
  title: string;
  /** @nullable */
  caption?: string | null;
  featured?: boolean;
}

export interface GalleryItemInput {
  imageUrl: string;
  category: string;
  title: string;
  caption?: string;
  featured?: boolean;
}

export interface GalleryItemUpdate {
  imageUrl?: string;
  category?: string;
  title?: string;
  caption?: string;
  featured?: boolean;
}

export interface Testimonial {
  id: number;
  name: string;
  rating: number;
  text: string;
  service: string;
  /** @nullable */
  avatarUrl?: string | null;
  featured?: boolean;
  createdAt?: string;
}

export interface TestimonialInput {
  name: string;
  rating: number;
  text: string;
  service: string;
  avatarUrl?: string;
  featured?: boolean;
}

export interface TestimonialUpdate {
  name?: string;
  rating?: number;
  text?: string;
  service?: string;
  avatarUrl?: string;
  featured?: boolean;
}

export interface StaffMember {
  id: number;
  name: string;
  role: string;
  experience: string;
  specializations: string[];
  languages?: string[];
  /** @nullable */
  imageUrl?: string | null;
  /** @nullable */
  bio?: string | null;
}

export interface StaffMemberInput {
  name: string;
  role: string;
  experience: string;
  specializations: string[];
  languages?: string[];
  imageUrl?: string;
  bio?: string;
}

export interface StaffMemberUpdate {
  name?: string;
  role?: string;
  experience?: string;
  specializations?: string[];
  languages?: string[];
  imageUrl?: string;
  bio?: string;
}

export interface ContactMessage {
  id: number;
  name: string;
  phone: string;
  /** @nullable */
  email?: string | null;
  /** @nullable */
  service?: string | null;
  message: string;
  read?: boolean;
  createdAt: string;
}

export interface ContactMessageInput {
  name: string;
  phone: string;
  email?: string;
  service?: string;
  message: string;
}

export interface ChatMessageInput {
  message: string;
}

export interface ChatResponse {
  reply: string;
  suggestions: string[];
  services?: string[];
}