# .migration-backup/lib/api-zod/src/generated/types/appointment.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * Mr Farrukh Hair Saloon API
 * OpenAPI spec version: 0.1.0
 */

export interface Appointment {
  id: number;
  name: string;
  phone: string;
  /** @nullable */
  email?: string | null;
  service: string;
  /** @nullable */
  stylist?: string | null;
  date: string;
  time: string;
  /** @nullable */
  notes?: string | null;
  status: string;
  createdAt: string;
}