# .migration-backup/lib/api-zod/src/generated/types/chatResponse.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * Mr Farrukh Hair Saloon API
 * OpenAPI spec version: 0.1.0
 */

export interface ChatResponse {
  reply: string;
  suggestions: string[];
  services?: string[];
}