# .migration-backup/lib/api-zod/src/generated/types/galleryItem.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * Mr Farrukh Hair Saloon API
 * OpenAPI spec version: 0.1.0
 */

export interface GalleryItem {
  id: number;
  imageUrl: string;
  category: string;
  title: string;
  /** @nullable */
  caption?: string | null;
  featured?: boolean;
}