# .migration-backup/lib/api-zod/src/generated/types/serviceInput.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * Mr Farrukh Hair Saloon API
 * OpenAPI spec version: 0.1.0
 */

export interface ServiceInput {
  name: string;
  category: string;
  description: string;
  imageUrl?: string;
  duration?: string;
  featured?: boolean;
}