# .migration-backup/lib/api-zod/src/generated/types/staffMember.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * Mr Farrukh Hair Saloon API
 * OpenAPI spec version: 0.1.0
 */

export interface StaffMember {
  id: number;
  name: string;
  role: string;
  experience: string;
  specializations: string[];
  languages?: string[];
  /** @nullable */
  imageUrl?: string | null;
  /** @nullable */
  bio?: string | null;
}