# .migration-backup/lib/api-zod/src/generated/types/staffMemberInput.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * Mr Farrukh Hair Saloon API
 * OpenAPI spec version: 0.1.0
 */

export interface StaffMemberInput {
  name: string;
  role: string;
  experience: string;
  specializations: string[];
  languages?: string[];
  imageUrl?: string;
  bio?: string;
}